+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.46.0] - 2026-02-14
+
+### Added
+
+- The `TypeMap` utility to convert from `NativeType` to its Typescript equivalent.
+
+### Changed
+
+- Renamed `Index` for `KeyDefinition`, `IndexAttributeDefinition` for `KeyAttributeDefinition`, and
+`IndexAttributeType` for `KeyAttributeType`.
+
 ## [0.45.0] - 2026-02-13
 
 ### Changed
@@ -591,6 +602,7 @@ intuitive.
 - Initial release of the package! Move the implementation work in progress from another
 project to here.
 
+[0.46.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.45.0...v0.46.0
 [0.45.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.44.0...v0.45.0
 [0.44.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.43.0...v0.44.0
 [0.43.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.42.1...v0.43.0